requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-decouple
# for pytest
httpx